Flask App structured and running
