+++ run.py
@@ -1,0 +1,8 @@
+from application import app
+
+if __name__ == '__main__':
+    #search(str(2015))
+    app.run(debug=True)
+    
+    #db.create_all()
+    #db.drop_all()
